+++ projects/sds/src/app.routes.ts
@@ -4,6 +4,10 @@ import {URL_CONST} from './app/shared/url/url.const';
 export const APP_ROUTES: Routes = [
 	{path: '', redirectTo: 'introductions/welcome', pathMatch: 'full'},
 	{
+		path: `introductions/newsletter`,
+		loadChildren: () => import('./app/component-page/component-pages.routes')
+	},
+	{
 		path: `introductions/:${URL_CONST.urlParams.selectedSlug}`,
 		loadChildren: () => import('./app/text-page/text-pages.routes')
 	},